refactor: Move enums to enums.ts
