+++ source/gallery.ts
@@ -1,7 +1,8 @@
 import type { Hitomi } from './hitomi';
 import { Image, Video } from './media';
 import { Language, Tag } from './tag';
-import { RESOURCE_DOMAIN, DEDICATED_TAG_PROPERTIES, SortType } from './utilities/constants';
+import { RESOURCE_DOMAIN, DEDICATED_TAG_PROPERTIES } from './utilities/constants';
+import { SortType } from './enums';
 import { defineProperties, hashTerm, formatOneOfState, toString } from './utilities/functions';
 import { Base, IndexProvider, HitomiError } from './utilities/structures';
 import type { URL, Node } from './utilities/types';
