+++ source/tag.ts
@@ -1,7 +1,8 @@
 import type { Hitomi } from './hitomi';
 import { Base, HitomiError } from './utilities/structures';
 import { formatOneOfState, hashTerm, toString } from './utilities/functions';
-import { BINARY_ORDERED_LANGUAGES, NameInitial, GALLERY_TYPES, LANGUAGE_NAMES, FRONT_DOMAIN, TAG_INDEX_DOMAIN, TAG_TYPES } from './utilities/constants';
+import { BINARY_ORDERED_LANGUAGES, GALLERY_TYPES, LANGUAGE_NAMES, FRONT_DOMAIN, TAG_INDEX_DOMAIN, TAG_TYPES } from './utilities/constants';
+import { NameInitial } from './enums';
 import type { Node } from './utilities/types';
 import type { Gallery } from './gallery';
